+++ UnrealDemo/Source/UnrealDemo/AuthenticateGameMode.h
@@ -1,0 +1,17 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+#pragma once
+
+#include "CoreMinimal.h"
+#include "GameFramework/GameMode.h"
+#include "AuthenticateGameMode.generated.h"
+
+/**
+ * 
+ */
+UCLASS()
+class UNREALDEMO_API AAuthenticateGameMode : public AGameMode
+{
+	GENERATED_BODY()
+	
+};