+++ UnrealDemo/Source/UnrealDemo/AuthenticatePlayerController.h
@@ -1,0 +1,18 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+#pragma once
+
+#include "CoreMinimal.h"
+#include "GameFramework/PlayerController.h"
+#include "AuthenticatePlayerController.generated.h"
+
+/**
+ * 
+ */
+UCLASS()
+class UNREALDEMO_API AAuthenticatePlayerController : public APlayerController
+{
+	GENERATED_BODY()
+
+	
+};