+++ UnrealDemo/Source/UnrealDemo/TcpClientWorker.h
@@ -1,10 +1,13 @@
 // Fill out your copyright notice in the Description page of Project Settings.
 
 #pragma once
-#include "Runnable.h"
-
+#include "Queue.h"
 #include "CoreMinimal.h"
 #include "Sockets.h"
+#include "BinarySerializer.h"
+#include "BinaryDeSerializer.h"
+#include "TcpCommandProcessor.h"
+#include "TcpClientSender.h"
 
 /**
  * 
@@ -12,13 +15,34 @@
 class UNREALDEMO_API TcpClientWorker : public FRunnable
 {
 public:
-	TcpClientWorker(TSharedPtr<FSocket> Socket) { this->Socket = Socket; };
+	TcpClientWorker(TSharedPtr<FSocket, ESPMode::ThreadSafe> Socket);
+	UFUNCTION()
 	void FlushAndComplete() { ExecuteLoop = false; };
 	virtual bool Init() override;
 	virtual uint32 Run() override;
 	virtual void Stop() override;
+	
+	UFUNCTION()
+	void OnTcpClientData(int32 BytesSent);
 
 private:
+	UPROPERTY()
 	bool ExecuteLoop = false;
-	TSharedPtr<FSocket> Socket;
+	UPROPERTY()
+	TWeakPtr<FSocket, ESPMode::ThreadSafe> Socket;
+	UPROPERTY()
+	UBinarySerializer* Serializer;
+	UPROPERTY()
+	UBinaryDeSerializer* DeSerializer;
+	UPROPERTY()
+	FRunnableThread* SendThread = nullptr;
+	UPROPERTY()
+	FRunnableThread* CmdProcessThread = nullptr;
+	UPROPERTY()
+	TcpCommandProcessor* CommandProcessorPtr = nullptr;
+	UPROPERTY()
+	TcpClientSender* ClientSenderPtr = nullptr;
+	UPROPERTY()
+	FDateTime LastActivity;
+	
 };