+++ UnrealDemo/Source/UnrealDemo/UnrealDemoConnectionStatus.h
@@ -4,7 +4,8 @@ enum class EGameServiceConnectionStatus: uint8{
 	CSTATUS_CONNECTION_ERROR UMETA(DisplayName = "Connection Error"),
 	CSTATUS_NOT_INITALIZED UMETA(DisplayName =  "Not Initialized"),
 	CSTATUS_NOT_CONNECTED UMETA(DisplayName = "Not Connected"),
-	CSTATUS_CONNECTED_TO_SERVER UMETA(DisplayName = "Connected to Server")
+	CSTATUS_CONNECTED_TO_SERVER UMETA(DisplayName = "Connected to Server"),
+	CSTATUS_UNKNOWN UMETA(DisplayName = "Unknown Problem")
 
 };