+++ CHANGELOG.md
@@ -2,3 +2,6 @@
 
 # Version 0.1.0
 * Build UE4 project as client / server project
+* Get connection to GameService
+* Send Heartbeats to GameService
+* Parse Binary from GameService